requirements.txt
requests
pytest
structlog
uuid
curlify
pydantic
pyhamcrest